added the code and data to find initial input control points
Data in forcpt.dat is the points along the curve for the len10400x jammed sweep. The python code does a terrible arclength parameterization and then fits polynomials to each coordinate dimension. The polyfits are then fed into Julia to get the control pts. The three lines of julia code are incuded in the header comment of the python script.